Improved container navigation, bmp and gif support
